pyproject.toml
[tool.poetry]
name = "emoji_bot"
version = "0.1.0"
description = "Reacts with relevant emoji"
authors = ["Tomer Keren <tomer.keren.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
"discord.py" = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
